repositories/AttackVLA/Pi0-Fast/robogcg_pi0_fast/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robogcg"
version = "0.1.0"
authors = [
    { name="Your Name", email="your.email@example.com" },
]
description = "A package for adversarial attacks on vision-language robot control models"
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "transformers>=4.4",
    "pillow>=9.0.0",
    "numpy>=1.20.0",
    "tqdm>=4.62.0",
    "matplotlib>=3.4.0",
]

[tool.setuptools]
packages = ["roboGCG"]

[project.urls]
Homepage = "https://github.com/yourusername/roboGCG"
Issues = "https://github.com/yourusername/roboGCG/issues"